requirements.txt
# Core LangChain packages
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.75
langchain-openai==0.3.27
langchain-pinecone==0.2.11
langchain-docling==1.0.0

# Streamlit Web App
streamlit==1.44.0

# Azure Blob Storage
azure-storage-blob==12.26.0
azure-core==1.35.0

# Pinecone Vector Database
pinecone==6.0.2

# OpenAI
openai==1.86.0

# Document Processing
unstructured[pdf]==0.17.2
pypdf==5.6.0
pdfminer.six==20250506
pikepdf==9.10.2
pdf2image==1.17.0

# Text Processing
tiktoken==0.9.0
nltk==3.9.1

# Environment and Configuration
python-dotenv==1.1.0
pydantic==2.11.6
pydantic-settings==2.9.1

# HTTP and API
httpx==0.28.1
requests==2.32.4
aiohttp==3.12.13

# Data Processing
numpy==2.2.6
pandas==2.3.0

# Machine Learning (for unstructured)
torch==2.8.0
torchvision==0.23.0
transformers==4.52.4
accelerate==1.10.1
timm==1.0.19

# Computer Vision (for PDF processing)
opencv-python==4.12.0.88
Pillow==11.2.1

# Google Cloud Vision (for unstructured)
google-cloud-vision==3.10.2

# ONNX Runtime (for unstructured)
onnx==1.19.0
onnxruntime==1.22.0

# Other dependencies
beautifulsoup4==4.13.4
lxml==5.4.0
chardet==5.2.0
filetype==1.2.0
python-magic==0.4.27
rapidfuzz==3.13.0
backoff==2.2.1
tqdm==4.67.1
psutil==7.0.0
python-oxmsg==0.0.2
html5lib==1.1
python-iso639==2025.2.18
langdetect==1.0.9
dataclasses-json==0.6.7
marshmallow==3.26.1
typing-inspect==0.9.0
mypy-extensions==1.1.0
emoji==2.14.1
python-multipart==0.0.20
huggingface-hub==0.34.4
matplotlib==3.10.3
scipy==1.15.3
pypdfium2==4.30.0
filelock==3.18.0
pyyaml==6.0.2
regex==2024.11.6
tokenizers==0.21.1
safetensors==0.5.3
fsspec==2025.5.1
hf-xet==1.1.3
protobuf==5.29.5
ml-dtypes==0.5.3
coloredlogs==15.0.1
flatbuffers==25.2.10
packaging==24.2
sympy==1.14.0
requests-toolbelt==1.0.0
zstandard==0.23.0
anyio==4.9.0
certifi==2025.4.26
httpcore==1.0.9
idna==3.10
h11==0.16
annotated-types==0.7.0
pydantic-core==2.33.2
typing-inspection==0.4.1
sniffio==1.3.1
charset-normalizer==3.4.2
urllib3==2.4.0
six==1.17.0
wrapt==1.17.2
py-cpuinfo==9.0.0
cffi==1.17.1
rich==14.0.0
pycparser==2.22
markdown-it-py==3.0.0
mdurl==0.1.2
distro==1.9.0
jiter==0.10.0
orjson==3.10.18
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.7.0
multidict==6.4.4
propcache==0.3.2
yarl==1.20.1
aiohttp-retry==2.9.1
iniconfig==2.1.0
pluggy==1.6.0
pygments==2.19.1
soupsieve==2.7
humanfriendly==10.0
pyasn1-modules==0.4.2
rsa==4.9.1
webencodings==0.5.1
MarkupSafe==3.0.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
kiwisolver==1.4.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
click==8.2.1
joblib==1.5.1
pytz==2025.2
tzdata==2025.2
cryptography==45.0.4
Deprecated==1.2.18
olefile==0.47
aiofiles==24.1.0
nest-asyncio==1.6.0
pycocotools==2.0.10
omegaconf==2.3.0
antlr4-python3-runtime==4.9.3
google-api-core==2.25.1
proto-plus==1.26.1
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.2
cachetools==5.5.2
pyasn1==0.6.1
pi-heif==1.1.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
effdet==0.4.1
unstructured-client==0.36.0
langsmith==0.3.45
tenacity==9.1.2
jsonpatch==1.33
jsonpointer==3.0.0